perf: add ENCODING_UTF8 constant and refactor Response class
This commit adds a new constant `ENCODING_UTF8` to the `common.ts` file.
It also refactors the `Response` class in the `Response.ts` file
by replacing all instances of `"utf-8"` with this new constant.
Additionally, it replaces all private method calls to `#send()` with `$end()`.
Finally, it updates some comments for clarity.