feat(config): Add support for view configuration
This commit adds support for configuring views in the Foxify framework. It introduces a new `ViewConfig` class that allows users to specify the directory containing view templates, the view template file extension, and a custom renderer function. The `Foxify` class now has methods to enable/disable view rendering and set the view engine.

The changes also include updates to the `Server` class to pass the view configuration settings when creating instances of `Request` and `Response`.